Added DockerClient and Payloads mixin
DockerClient exposes a subset of the docker API as described by
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.21/
JSON results are marshalled to OpenStruct DTOs by the methods in the
Payloads mixin.

Editorial choice made to not modify the payload attribute names to
be rubyish.  This means attributes like "ImageID" are accessed in
the DTOs via dto.ImageID and not dto.image_id.  While this deviates
from ruby semantics a bit, it mirrors the docker API documentation
which will have API results described as:

{
  "ImageID: "asdasid930ekad9013easd"
}

I would rather not have the cognitive overhead of translating the
documented API atttributes to something rubyish and vice versa
from here to eternity.